ImageWatermark/ImageWatermark/requirements.txt
Flask==1.1.2
numpy==1.18.5
Pillow==8.0.1
opencv-python-headless==4.2.0.32
gunicorn